requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
loguru>=0.7.0

# API clients
requests>=2.31.0
yfinance>=0.2.18
ccxt>=4.1.0

# Sentiment Analysis
textblob>=0.17.1
vaderSentiment>=3.3.2

# Scientific Computing
scipy>=1.10.0
scikit-learn>=1.3.0

# Optional ML (for alpha detection)
xgboost>=2.0.0
lightgbm>=4.0.0

# Data feeds
feedparser>=6.0.10
websockets>=12.0
